refacto: use combine to prevent pulling from data
Use combine to automatically propagate changes of the repository to the statistics view model instead of calling the load method at each appear of the view
There is a small workaround to be able to still use the @Published property wrapper and also protocol for the repository but the advantage of not having to pull the data is higher.
It also prevent a bug linked to onAppear not being called at the first appearance of the SwiftUI tab.